create utils submodule and move util and latent_algebra
