+++ cleanlab/__init__.py
@@ -1,8 +1,7 @@
 from .version import __version__
 from . import classification
-from . import latent_algebra
 from . import count
 from . import rank
 from . import filter
 from . import noise_generation
-from . import util
+from .utils import util
