+++ cleanlab/utils/latent_algebra.py
@@ -26,7 +26,7 @@
 
 import numpy as np
 
-from cleanlab.util import value_counts, clip_values, clip_noise_rates
+from cleanlab.utils.util import value_counts, clip_values, clip_noise_rates
 import warnings
