+++ tests/test_latent_algebra.py
@@ -14,7 +14,7 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with cleanlab.  If not, see <https://www.gnu.org/licenses/>.
 
-from cleanlab import latent_algebra
+from cleanlab.utils import latent_algebra
 import numpy as np
 import pytest
